fix incorrect link
When trying to download a deb via a device that's not through cydia, the page returns the default repo welcome page. Since the directory is /debs, the admin link goes to http://site/debs/Admin.php when it should be http://site/Admin.php. This little fix corrects this issue.